Support multi-line string values for openapi keys
If the openapi yaml description contains multi-line strings (for instance a nice descriptive 'description' that is written as block text), then using a single-quoted string for this value results in an error when black tries to process the file. Under this example of a long description, you might get something like:

```
black.InvalidInput: Cannot parse: 23:17:     description = "
```

Using triple-quoted strings avoids this.